libtego: reduced file and message id types to 32 bits and added static assert verifying tego types match protocol type

+++ src/libtego/include/tego/tego.h
@@ -820,9 +820,9 @@ void tego_context_get_tor_bootstrap_status(
 // milliseconds since 1970-01-01T00:00:00 utc.
 typedef uint64_t tego_time_t;
 // unique (per user) message identifier
-typedef uint64_t tego_message_id_t;
+typedef uint32_t tego_message_id_t;
 // unique (per user) attachment identifier
-typedef uint64_t tego_attachment_id_t;
+typedef uint32_t tego_attachment_id_t;
 // struct for file hash
 typedef struct tego_file_hash tego_file_hash_t;
